+++ Source/RxSwift/Tron+RxSwift.swift
@@ -49,7 +49,7 @@ extension APIRequest {
     }
 }
 
-extension MultipartAPIRequest {
+extension UploadAPIRequest {
     /**
      Creates an Observable<Model> for multipart upload.
      
@@ -57,7 +57,8 @@ extension MultipartAPIRequest {
      
      - returns: Observable<Model>
      */
-    open func rxMultipartResult(memoryThreshold threshold: UInt64 = SessionManager.multipartFormDataEncodingMemoryThreshold) -> Observable<Model> {
+    open func rxMultipartResult(memoryThreshold threshold: UInt64 = SessionManager.multipartFormDataEncodingMemoryThreshold,
+                                uploadProgress: (Request.ProgressHandler)? = nil) -> Observable<Model> {
         return Observable.create { observer in
             var request : Alamofire.Request?
             self.performMultipart(withSuccess: { result in
@@ -70,6 +71,9 @@ extension MultipartAPIRequest {
                 encodingCompletion : { completion in
                     if case let SessionManager.MultipartFormDataEncodingResult.success(originalRequest, _, _) = completion {
                         request = originalRequest
+                        if let progressClosure = uploadProgress {
+                            originalRequest.uploadProgress(closure: progressClosure)
+                        }
                     }
             })
             return Disposables.create {